+++ backend/src/test/java/edonymyeon/backend/member/integration/MemberConsumptionIntegrationTest.java
@@ -1,8 +1,8 @@
 package edonymyeon.backend.member.integration;
 
-import static edonymyeon.backend.member.integration.memberConsumptionSteps.MemberConsumptionSteps.구매_확정_요청을_보낸다;
-import static edonymyeon.backend.member.integration.memberConsumptionSteps.MemberConsumptionSteps.절약_확정_요청을_보낸다;
-import static edonymyeon.backend.member.integration.memberConsumptionSteps.MemberConsumptionSteps.확정_취소_요청을_보낸다;
+import static edonymyeon.backend.member.integration.steps.MemberConsumptionSteps.구매_확정_요청을_보낸다;
+import static edonymyeon.backend.member.integration.steps.MemberConsumptionSteps.절약_확정_요청을_보낸다;
+import static edonymyeon.backend.member.integration.steps.MemberConsumptionSteps.확정_취소_요청을_보낸다;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import edonymyeon.backend.IntegrationTest;
